Add options to externalize showcase storage
Adds two callback functions to Showcase widget:
 * isSlideVisible
 * hideSlide

That allows to externalize slide visibility and control it from consumer
application code.

Both arguments have default values that stores data in
SharedPreferences, to keep the same default behaviour of library.

Main driver for this change is to improve end user experience when app
is used on multiple devices. In such case, state of slides may be stored
on the user profile and then ignored when app is started on a different
device.